+++ src/functions/replaceSpaces.js
@@ -6,7 +6,10 @@
  * @param {string} space _Optional:_ The space to replace it with. Defaults to a one-width space (` `).
  * @returns
  */
-const _replaceSpaces = (text, spaceChar = "/s", space = " ") =>
-  String(text).replace(new RegExp(spaceChar, "ig"), space);
+const _replaceSpaces = (text, spaceChar = "/s", space = " ") => {
+  if (typeof text !== "undefined") {
+    return text.replace(new RegExp(spaceChar, "ig"), space);
+  }
+};
 
 module.exports = _replaceSpaces;